API change - move and rename delegates in class Pcap, PacketArrivalEvent, PcapStatisticsModeEvent and PcapCaptureStoppedEvent, to be directly in the SharpPcap namespace
Move and rename PacketArrivalEvent to PacketArrivalEventHandler in a file of the same name
Move and rename PcapStatisticsModeEvent to StatisticsModeEventHandler in a file of the same name
Move and rename PcapCaptureStoppedEvent to CaptureStoppedEventHandler in a file of the same name.

This change moves the delegates from outside of the SharpPcap.Pcap class to
be directly in the SharpPcap namespace. For example, SharpPcap.Pcap.PacketArrivalEvent to SharpPcap.PacketArrivalEventHandler.

There was no need to nest these delegates and it makes them more difficult to use as it requires specifying another class.

Fix examples to work with these API changes