Fix receiving CaptureEvents from OfflinePcapDevices by using the base type, PcapDevice, as the internal device type. Also fix a bug preventing packet capturing with Mono.Posix and offline files.
To help users with calling device specific methods like Statistics() I had made the type
of the CaptureEventArgs.Device property a LivePcapDevice. This breaks things with
OfflinePcapDevices though because the type is assumed to be a LivePcapDevice.

To fix things the type of CaptureEventArgs.Device was switched to PcapDevice. A abstract
method Statistics() was added to PcapDevice. OfflinePcapDevice implements this method and
throws a NotSupportedOnOfflineDeviceException. Users can call the Statistics() method of
CaptureEventArgs.Device without casting although it presumes that the device is a
LivePcapDevice.

Added OfflinePcapDeviceTest.cs that tests capturing packets from offline devices in the same
way as live devices. It also verifies that the appropriate exception is thrown from
OfflinePcapDevice.Statistics()

Fix a bug in PcapDeviceCaptureLoop.cs with offline devices and libPcap and Mono.Posix.
Calling poll() on the file descriptor of the offline file isn't going to produce the
events that a live device would produce so packets are never received. Simply avoid
using the Mono.Posix code unless we are a LivePcapDevice.