Fix PosixTimeval.ToString() when microseconds is < 100000
Bug would cause PosixTimeval.ToString() with a seconds of 123 and a microseconds of 1234 to return "123.1234"
but this should be returning "123.001234" (note the leading zeros).

Change from StringBuilder.Append() to StringBuilder.AppendFormat() and add a format that will add leading '0's
until there are 6 digits of microseconds.

Bug reported by github user bucweat.