Use Span<> to move memory allocation out of SharpPcap to remove memory allocation in high performance use cases
Goal is to improve the highest performance use cases where returning RawCapture
is a performance bottleneck due to memory allocation, while retaining ease of use.

PcapDevice.GetNextPacket() retains interfaces that return RawCapture as before but capture loops
assume the usage of Span<>.

Modify CaptureEventArgs to be a ref struct and pass in a Span<byte>. Modify CaptureEventArgs.Packet
to create and cache a RawCapture instance on the getter.

Modify WinDivertDevice to match the updated PacketArrivalEventHandler.