Consolidate synchronous packet retrival into 'int GetNextPacket(out CaptureEventArgs e)' to simplify API.
Consolidation enables removing duplicate code in PcapDevice and WinDivertDevice

This is also intended to simplify the API for the end user. Now all calls to GetNextPacket()
make use of Span<> for highest performance, and CaptureEventArgs.Packet converts to a RawPacket
if that property is retrieved.

Note: This is an api breaking change as the return value from GetNextPacket() is now 1
upon success, 0 if no error and negative upon failure. Previously it was 1 upon success
for the packet retrieval functions that returned a status.